requirements.txt
asgiref==3.8.1
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.1.4
django-admin-sortable2==2.2.8
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-extensions==4.1
django-jazzmin==3.0.1
djangorestframework==3.15.2
graphviz==0.20.3
idna==3.10
pillow==11.1.0
qrcode==8.2
requests==2.32.3
sqlparse==0.5.2
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
